Garmin: fix basetype handling and truncation
Numeric BaseTypes now return fractional part if a scale is set, the return type for numeric values is unboxed from Number class so that existing comparisons and checks keep working, but values are not truncated at nearest integer value.

The codegen class has been updated and some of the messages are re-generated with the new code, with the exception of Hrv* messages, since updating those needs updating the related sample types.

Test cases expected values have been adjusted to keep into account the fractional part of some fields.

Garmin: Fix HRV value and summary types

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/fit/baseTypes/BaseTypeFloat.java
@@ -26,7 +26,7 @@ public class BaseTypeFloat implements BaseTypeInterface {
         }
         if (Float.isNaN(f) || f == invalid)
             return null;
-        return (float) (f / scale) - offset;
+        return (f / scale) - offset;
     }
 
     @Override
